feat(http-resources): move firebase methods to ducks module
