--- web-client/src/http/resources/auth/auth.ts
@@ -1,16 +1,0 @@
-import { AxiosRequestConfig } from 'axios';
-import firebase, { firebaseAuth } from 'src/firebase';
-import { IHTTPRequest } from 'src/http/HTTPRequest';
-
-export const login = (request: IHTTPRequest) => {
-  const config: AxiosRequestConfig = {
-    method: 'POST',
-    url: '/login',
-  };
-  return request.execute(config).then((response: any) => response.data);
-};
-
-export const observeUser = (nextValue: Function): firebase.Unsubscribe =>
-  firebaseAuth.onAuthStateChanged((user: firebase.User | null) => {
-    nextValue(user);
-  });
